Fixed gamelobby error
Error: If there is only one player in the lobby, the player can start alone.

Fixed: If there are not two players in the lobby, the game does not start although the only one player in the lobby is in ready status. (= If there are not two players in the lobby, the server does not send 'start game' event to client.)